implemented float but cant with gantt

+++ Ass/src/assig/FCFS.java
@@ -16,8 +16,8 @@ public class FCFS extends Frame{
    void btnAddPressed() {
        //System.out.println("in FCFS");
         Process p = new Process();
-        p.setBurst(Integer.valueOf(txtBurst.getText()));
-        p.setArrival(Integer.valueOf(txtArrival.getText()));
+        p.setBurst(Float.valueOf(txtBurst.getText()));
+        p.setArrival(Float.valueOf(txtArrival.getText()));
         p.setName(count++);
         //set priority of each process to be arrival time
        // p.setPriority(Integer.valueOf(txtArrival.getText()));
